refactor(canvas): extract chat messages controller
