+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatMessages/index.tsx
@@ -1,12 +1,4 @@
-import {
-  useCallback,
-  useEffect,
-  useLayoutEffect,
-  useRef,
-  useState,
-  type KeyboardEvent,
-  type MouseEvent,
-} from 'react';
+import type { KeyboardEvent } from 'react';
 import './index.css';
 import type { AgentChatMessage, CanvasNode } from '../../../types';
 import { BotAvatarIcon } from '../../icons';
@@ -16,42 +8,8 @@ import { ChatThreadSkeleton } from './ChatThreadSkeleton';
 import type { PendingClarification, ToolCallStatus } from '../../../types';
 import { buildAnchorElementId } from '../utils/anchors';
 import { useI18n } from '../../../i18n';
-import { isVSCodeLink } from '../utils/externalLinks';
-import { localPathFromHref } from '../utils/localFileLinks';
 import { ChatClarificationCard } from './ChatClarificationCard';
-import { useChatMessagesStatus } from './useChatMessagesStatus';
-import { tabRefFromMentionElement } from '../utils/tabMentions';
-/** How close (px) to the bottom still counts as "reading the tail" — within
- *  this band the view keeps following the stream; beyond it the user has
- *  scrolled up to read and auto-scroll must not yank them back. */
-const PIN_THRESHOLD_PX = 80;
-const CONVERSATION_SCROLL_CACHE_LIMIT = 50;
-// Dock workspace switching owns a 2.5s acknowledgement window. Chat waits a
-// little longer so a legitimate late switch cannot be mislabeled as stale.
-const TAB_ACTIVATION_TIMEOUT_MS = 3000;
-const TAB_ACTIVATION_SUCCESS_FEEDBACK_MS = 1600;
-
-type DockTabActivationResult = { status: 'activated' | 'reopened' | 'stale' };
-
-interface DockTabActivationRequestDetail {
-  tabId: string;
-  dockWorkspaceId?: string;
-  tab?: ReturnType<typeof tabRefFromMentionElement>;
-  respond: (result: DockTabActivationResult) => void;
-}
-
-// Session switches remount the transcript content, not the scroll container.
-// Keep a small LRU-like cache so returning to a conversation restores the
-// reader's place without retaining an unbounded session history.
-const conversationScrollPositions = new Map<string, number>();
-
-const rememberConversationScroll = (key: string, scrollTop: number): void => {
-  conversationScrollPositions.delete(key);
-  conversationScrollPositions.set(key, scrollTop);
-  if (conversationScrollPositions.size <= CONVERSATION_SCROLL_CACHE_LIMIT) return;
-  const oldestKey = conversationScrollPositions.keys().next().value;
-  if (oldestKey) conversationScrollPositions.delete(oldestKey);
-};
+import { useChatMessagesController } from './useChatMessagesController';
 
 interface ChatMessagesProps {
   messages: AgentChatMessage[];
@@ -144,259 +102,28 @@ export const ChatMessages = ({
   conversationKey,
 }: ChatMessagesProps) => {
   const { t } = useI18n();
-  const messagesEndRef = useRef<HTMLDivElement>(null);
-  const containerRef = useRef<HTMLDivElement>(null);
-  // Whether the viewport is glued to the newest message. Ref drives the
-  // scroll-follow decision synchronously; state mirrors it for the
-  // "jump to latest" affordance.
-  const pinnedRef = useRef(true);
-  const [atBottom, setAtBottom] = useState(true);
-  const [tabNavigationFeedback, setTabNavigationFeedback] = useState<{
-    message: string;
-    tone: 'progress' | 'success' | 'error';
-  } | null>(null);
-  const tabNavigationFeedbackTimerRef = useRef<number>();
-  const tabNavigationRequestRef = useRef(0);
-  const prevCountRef = useRef(0);
-  const prevSessionLoadingRef = useRef(sessionLoading);
-  const previousConversationKeyRef = useRef(conversationKey);
-  const restoredConversationKeyRef = useRef<string>();
-  const skipNextFollowEffectRef = useRef(false);
-  // While a smooth programmatic scroll glides down, intermediate scroll
-  // events report "not at bottom" — ignore them briefly so the jump button
-  // doesn't flash mid-animation.
-  const autoScrollUntilRef = useRef(0);
-  const latestTurnStatus = messages[messages.length - 1]?.turnStatus;
-  const { skeletonVisible, turnAnnouncement } = useChatMessagesStatus({
-    latestTurnStatus,
+  const {
+    atBottom,
+    containerRef,
+    handleMessageClick,
+    handleScroll,
+    messagesEndRef,
+    scrollToLatest,
+    skeletonVisible,
+    tabNavigationFeedback,
+    turnAnnouncement,
+  } = useChatMessagesController({
+    messages,
     loading,
-    messageCount: messages.length,
+    streamingTools,
+    pendingClarify,
     pendingLabel,
     sessionLoading,
+    conversationKey,
+    interactionDisabled,
+    onSessionJump,
+    onNodeFocus,
   });
-
-  useEffect(() => () => {
-    tabNavigationRequestRef.current += 1;
-    if (tabNavigationFeedbackTimerRef.current !== undefined) {
-      window.clearTimeout(tabNavigationFeedbackTimerRef.current);
-    }
-  }, []);
-
-  const handleScroll = useCallback(() => {
-    const el = containerRef.current;
-    if (!el) return;
-    if (conversationKey) rememberConversationScroll(conversationKey, el.scrollTop);
-    const distance = el.scrollHeight - el.scrollTop - el.clientHeight;
-    const pinned = distance < PIN_THRESHOLD_PX;
-    if (!pinned && performance.now() < autoScrollUntilRef.current) return;
-    pinnedRef.current = pinned;
-    setAtBottom(pinned);
-  }, [conversationKey]);
-
-  const scrollToLatest = useCallback((behavior: ScrollBehavior) => {
-    const reducedMotionQuery = typeof window.matchMedia === 'function'
-      ? window.matchMedia('(prefers-reduced-motion: reduce)')
-      : undefined;
-    const reducedMotion = reducedMotionQuery?.matches ?? false;
-    const effectiveBehavior = behavior === 'smooth' && reducedMotion ? 'auto' : behavior;
-    autoScrollUntilRef.current = effectiveBehavior === 'smooth' ? performance.now() + 600 : 0;
-    pinnedRef.current = true;
-    setAtBottom(true);
-    messagesEndRef.current?.scrollIntoView({ behavior: effectiveBehavior });
-  }, []);
-
-  useLayoutEffect(() => {
-    const el = containerRef.current;
-    const previousKey = previousConversationKeyRef.current;
-    if (previousKey !== conversationKey) {
-      if (previousKey && el) rememberConversationScroll(previousKey, el.scrollTop);
-      previousConversationKeyRef.current = conversationKey;
-      restoredConversationKeyRef.current = undefined;
-      skipNextFollowEffectRef.current = true;
-    }
-    if (
-      !conversationKey
-      || sessionLoading
-      || restoredConversationKeyRef.current === conversationKey
-      || !el
-    ) {
-      return;
-    }
-
-    const savedScrollTop = conversationScrollPositions.get(conversationKey);
-    if (savedScrollTop === undefined) {
-      scrollToLatest('auto');
-    } else {
-      el.scrollTop = savedScrollTop;
-      const distance = el.scrollHeight - el.scrollTop - el.clientHeight;
-      const pinned = distance < PIN_THRESHOLD_PX;
-      pinnedRef.current = pinned;
-      setAtBottom(pinned);
-    }
-    restoredConversationKeyRef.current = conversationKey;
-    skipNextFollowEffectRef.current = true;
-  }, [conversationKey, scrollToLatest, sessionLoading]);
-
-  useEffect(() => {
-    const prevCount = prevCountRef.current;
-    prevCountRef.current = messages.length;
-    const sessionJustLoaded = prevSessionLoadingRef.current && !sessionLoading;
-    prevSessionLoadingRef.current = sessionLoading;
-    // A message the user just sent — or a fresh session load — always snaps
-    // the view to the bottom. Otherwise only follow the stream while the
-    // user is already reading the tail; never yank them back up-thread.
-    const lastIsUser = messages.length > 0 && messages[messages.length - 1].role === 'user';
-    const userJustSent = messages.length > prevCount && lastIsUser;
-    const sessionReset = messages.length < prevCount;
-    if (skipNextFollowEffectRef.current) {
-      skipNextFollowEffectRef.current = false;
-      return;
-    }
-    if (userJustSent || sessionReset || sessionJustLoaded) {
-      scrollToLatest(userJustSent ? 'smooth' : 'auto');
-      return;
-    }
-    if (pinnedRef.current) {
-      messagesEndRef.current?.scrollIntoView({ behavior: 'auto' });
-    }
-  }, [messages, pendingClarify, pendingLabel, sessionLoading, streamingTools, scrollToLatest]);
-
-  const handleMessageClick = useCallback(async (event: MouseEvent) => {
-    const target = event.target as HTMLElement | null;
-    if (!target) return;
-
-    // Copy-code button rendered by the markdown fence renderer.
-    const copyBtn = target.closest<HTMLButtonElement>('[data-action="copy-code"]');
-    if (copyBtn) {
-      const codeEl = copyBtn.closest('.chat-code-block')?.querySelector('code');
-      const code = (codeEl?.textContent ?? '').replace(/\n$/, '');
-      try {
-        await navigator.clipboard.writeText(code);
-        copyBtn.dataset.state = 'copied';
-        copyBtn.textContent = t('chat.copied');
-        window.setTimeout(() => {
-          delete copyBtn.dataset.state;
-          copyBtn.textContent = t('chat.copy');
-        }, 1200);
-      } catch {
-        /* clipboard unavailable — ignore */
-      }
-      return;
-    }
-
-    // VS Code protocol links from markdown should launch the editor through
-    // main-process shell.openExternal; letting target=_blank handle them asks
-    // Electron to create a BrowserWindow for a custom scheme.
-    const link = target.closest<HTMLAnchorElement>('a[href]');
-    const href = link?.getAttribute('href') ?? '';
-    const localPath = localPathFromHref(href);
-    if (localPath) {
-      event.preventDefault();
-      event.stopPropagation();
-      void window.canvasWorkspace.file.openPath(localPath);
-      return;
-    }
-    if (href && isVSCodeLink(href)) {
-      event.preventDefault();
-      event.stopPropagation();
-      void window.canvasWorkspace.shell.openExternal(href);
-      return;
-    }
-
-    // Session-ref chip → load session and scroll to the matched message.
-    // Blocked while a turn is streaming: switching sessions mid-generation
-    // would clobber the in-flight assistant message.
-    const sessionChip = target.closest<HTMLElement>('[data-action="session-jump"]');
-    if (sessionChip) {
-      if (loading || sessionLoading || interactionDisabled || !onSessionJump) return;
-      const sid = sessionChip.dataset.sessionId;
-      const wid = sessionChip.dataset.workspaceId;
-      const mi = sessionChip.dataset.messageIndex;
-      const parsedIndex = mi !== undefined && mi !== '' ? Number(mi) : undefined;
-      if (sid && wid) {
-        onSessionJump(sid, wid, Number.isInteger(parsedIndex) ? parsedIndex : undefined);
-      }
-      return;
-    }
-
-    // Tab mention chip → activate the referenced right-dock tab. Broadcast so
-    // the dock (which owns tab state) can switch to it; harmless where no dock
-    // is mounted (full-screen chat).
-    const tabChip = target.closest<HTMLElement>('[data-action="tab-jump"]');
-    if (tabChip) {
-      const tabId = tabChip.dataset.tabId;
-      if (!tabId) return;
-      const name = tabChip.querySelector('.chat-mention-chip-label')?.textContent?.trim() || tabId;
-      const unavailableMessage = t('chat.tabNavigation.unavailable', { name });
-      if (tabChip.dataset.state === 'stale') {
-        setTabNavigationFeedback({ message: unavailableMessage, tone: 'error' });
-        return;
-      }
-
-      if (tabNavigationFeedbackTimerRef.current !== undefined) {
-        window.clearTimeout(tabNavigationFeedbackTimerRef.current);
-      }
-      setTabNavigationFeedback({
-        message: t('chat.tabNavigation.opening', { name }),
-        tone: 'progress',
-      });
-      const request = ++tabNavigationRequestRef.current;
-      let settled = false;
-      const respond = (result: DockTabActivationResult) => {
-        if (settled || request !== tabNavigationRequestRef.current) return;
-        settled = true;
-        if (tabNavigationFeedbackTimerRef.current !== undefined) {
-          window.clearTimeout(tabNavigationFeedbackTimerRef.current);
-        }
-        if (result.status === 'activated' || result.status === 'reopened') {
-          setTabNavigationFeedback({
-            message: t(result.status === 'reopened'
-              ? 'chat.tabNavigation.reopened'
-              : 'chat.tabNavigation.opened', { name }),
-            tone: 'success',
-          });
-          tabNavigationFeedbackTimerRef.current = window.setTimeout(() => {
-            setTabNavigationFeedback(null);
-          }, TAB_ACTIVATION_SUCCESS_FEEDBACK_MS);
-          return;
-        }
-        tabChip.dataset.state = 'stale';
-        tabChip.classList.add('chat-mention-chip--stale');
-        tabChip.setAttribute('aria-disabled', 'true');
-        tabChip.title = unavailableMessage;
-        setTabNavigationFeedback({ message: unavailableMessage, tone: 'error' });
-      };
-      tabNavigationFeedbackTimerRef.current = window.setTimeout(
-        () => respond({ status: 'stale' }),
-        TAB_ACTIVATION_TIMEOUT_MS,
-      );
-      const detail: DockTabActivationRequestDetail = {
-        tabId,
-        dockWorkspaceId: tabChip.dataset.dockWorkspaceId,
-        tab: tabRefFromMentionElement(tabChip),
-        respond,
-      };
-      window.dispatchEvent(new CustomEvent('canvas:activate-dock-tab', { detail }));
-      return;
-    }
-
-    // File/folder mention chip → open the referenced project path in VS Code.
-    const fileChip = target.closest('[data-file-path]') as HTMLElement | null;
-    const filePath = fileChip?.dataset.filePath;
-    if (filePath) {
-      void window.canvasWorkspace.file.openInVSCode(filePath);
-      return;
-    }
-
-    // Mention chip → focus the canvas node it references.
-    const chip = target.closest('.chat-mention-chip--clickable') as HTMLElement | null;
-    if (!chip || !onNodeFocus) return;
-    const nodeId = chip.dataset.nodeId;
-    if (nodeId) {
-      onNodeFocus(nodeId);
-    }
-  }, [interactionDisabled, loading, onNodeFocus, onSessionJump, sessionLoading, t]);
-
   const hasStreamingAssistantMessage = loading
     && messages.length > 0
     && messages[messages.length - 1].role === 'assistant';
